+++ src/components/Navigation/FincNavigation.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+import { useStripes } from '@folio/stripes/core';
 
 import {
   Button,
@@ -12,11 +13,16 @@ import urls from '../DisplayUtils/urls';
 const FincNavigation = ({
   id,
 }) => {
+  const stripes = useStripes();
+  const hasCollectionsPerms = stripes.hasPerm('finc-config.metadata-collections.collection.get');
+  const hasSourcesPerms = stripes.hasPerm('finc-config.metadata-sources.collection.get');
+
   return (
     <ButtonGroup fullWidth data-test-navigation>
       <Button
         buttonStyle={id === 'source' ? 'primary' : 'default'}
         data-test-navigation-source
+        disabled={!hasSourcesPerms}
         id="metadata-sources"
         to={id === 'collection' ? urls.sources() : ''}
       >
@@ -25,6 +31,7 @@ const FincNavigation = ({
       <Button
         buttonStyle={id === 'collection' ? 'primary' : 'default'}
         data-test-navigation-collection
+        disabled={!hasCollectionsPerms}
         id="metadata-collections"
         to={id === 'source' ? urls.collections() : ''}
       >